lc: fix gcm buffer lengths
The Liquid Crypto API changed the condition for adding digest space:

OLD: if (op->digest != NULL)
         dlen += digest_len;

NEW: if ((op->digest != NULL) ||
         (op->digest == NULL && op->out_buffer != NULL && op->encrypt))
         dlen += digest_len;

This broke lc gcm implementation because:

1. Old code passed buffer length including tag (input_len + tls_tag_len)
   with digest=NULL

2. For encryption: NEW LC API adds digest_len due to new condition
   (digest==NULL && out_buffer!=NULL && encrypt), causing double-counting
   of tag space and buffer overruns

3. For decryption: NEW LC API expects explicit digest pointer for tag
   verification but digest was NULL, causing verification failures

Fix:
- Pass only data length (input_len) to prepare_lc_buf, excluding tag
- Set digest pointer explicitly to tag location:
  * Encryption: output_buf + input_len
  * Decryption: input_buf + input_len
- Let LC API add digest space based on digest pointer

JIRA: https://essjira.marvell.com/browse/IPBUSW-85843

+++ pal/liquid_crypto/defs.h
@@ -158,8 +158,8 @@ typedef struct pal_gcm_ctx {
 	struct dao_lc_sym_op *op;
 	uint8_t **input_buf;
 	uint8_t **output_buf;
-	long int *input_len;
-	long int *output_len;
+	size_t *input_len;
+	size_t *output_len;
 	char aad_pipe[SSL_MAX_PIPELINES][TLS_AAD_LEN];
 	uint8_t key[32];
 	uint64_t iv[3];
@@ -184,7 +184,7 @@ typedef struct pal_gcm_ctx {
 	async_job async_cb;
 } pal_gcm_ctx_t;
 
-int prepare_lc_buf(struct dao_lc_buf **head, uint8_t *data, long int len);
+int prepare_lc_buf(struct dao_lc_buf **head, uint8_t *data, size_t len);
 int sym_create_session(uint16_t dev_id,
 		struct dao_lc_sym_ctx cry_session,struct dao_lc_cmd_event *event, uint8_t reconfigure, uint64_t sess_cookie);
 int sym_session_cleanup(struct dao_lc_cmd_event *event, int dev_id);
